Refactored scene size to u32
While this means size() calculates the scaled size each time, it allows
for lossless paths to be used when dealing with pixels.

Additionally, clipping now is ensured to be constrained to the scene's
size, whereas previously it went unchecked. This only becomes an issue
if you have inaccuracies in floating point numbers in most situations.

+++ core/src/frame_renderer/frame.rs
@@ -13,7 +13,7 @@ use crate::{
 };
 #[derive(Default, Debug)]
 pub struct Frame {
-    pub size: Size<f32, ScreenSpace>,
+    pub size: Size<u32, ScreenSpace>,
     pub textures: HashMap<u64, Texture>,
     pub(crate) commands: Vec<FrameCommand>,
     pub(crate) fonts: HashMap<u64, LoadedFont>,
@@ -23,7 +23,7 @@ pub struct Frame {
 
 #[derive(Default, Debug)]
 struct FrameReceiver {
-    size: Size<f32, ScreenSpace>,
+    size: Size<u32, ScreenSpace>,
     elements: Vec<Element>,
 }
